+++ CHANGELOG.md
@@ -21,7 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [frontend] - Make human messages not span the entire width of the chat.
 - Copyright and acknowledgement.
 - pre-commit hooks for frontend.
-- Better system prompt.
+- [frontend] - Loading bar and better first message.
+- Better system prompt
 - Sort components in folder.
 - Brain region/cell hierarchies read from storage (not from local filesystem).
 - Authentication is done directly through keycloak.